Data storage now works! Added confirmation for closing the app, improved create poll process, and much more on the way!

+++ src/db/_db_structs.hpp
@@ -32,7 +32,7 @@ struct Poll
     // id GIVEN to users/voters to allow them to vote in the poll
     std::string name;
     std::string voter_id;
-    int owner_id;
+    size_t owner_id;
     std::time_t created_at;
 };
 
@@ -40,7 +40,7 @@ struct PollOption
 {
     size_t id;
 
-    int poll_id;
+    size_t poll_id;
     std::string name;
     // used in GUI, a brief description of the option, can be an empty string.
     std::string description;
